Step3: Low_res and High_res lists are working, with Add, Remove, SelectAll and DeselectAll buttons:
When an object is included in one of both lists the pertinent object variable (Jet.high_res or Jet.low_res) is set to True, and the other one is set to False.
When an object is removed both variables are set to False.
This strategy makes easier the selection or deselection of the objects in the lists.